+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxContentList/ComboboxContentList.tsx
@@ -1,12 +1,57 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
+import classnames from "classnames";
 import { Text } from "@jobber/components/Text";
 import styles from "./ComboboxContentList.css";
 import { ComboboxListProps } from "../../../Combobox.types";
 import { ComboboxOption } from "../../ComboboxOption/ComboboxOption";
 
 export function ComboboxContentList(props: ComboboxListProps): JSX.Element {
+  const [listScrollState, setlistScrollState] = useState("");
+
+  useEffect(() => {
+    const handleScroll = () => {
+      if (props.optionsListRef.current) {
+        const { scrollTop, clientHeight, scrollHeight } =
+          props.optionsListRef.current;
+
+        if (scrollHeight === clientHeight) {
+          setlistScrollState("scrollNone");
+        } else if (scrollTop === 0) {
+          setlistScrollState("scrollTop");
+        } else if (scrollTop + clientHeight === scrollHeight) {
+          setlistScrollState("scrollBottom");
+        } else {
+          setlistScrollState("");
+        }
+      }
+    };
+
+    if (!props.showEmptyState && props.options.length === 0) {
+      handleScroll();
+    }
+
+    if (props.optionsListRef.current) {
+      props.optionsListRef.current.addEventListener("scroll", handleScroll);
+      handleScroll();
+    }
+
+    return () => {
+      if (props.optionsListRef.current) {
+        props.optionsListRef.current.removeEventListener(
+          "scroll",
+          handleScroll,
+        );
+      }
+    };
+  }, [props.options]);
+
   return (
-    <div className={styles.container}>
+    <div
+      className={classnames(
+        styles.container,
+        styles[listScrollState as keyof typeof styles],
+      )}
+    >
       {!props.showEmptyState && props.options.length > 0 && (
         <ul
           className={styles.optionsList}