+++ packages/components/src/DataList/components/DataListAction/DataListAction.tsx
@@ -24,10 +24,20 @@ export function DataListAction<T extends DataListObject>({
 
   const color = destructive ? "critical" : "heading";
 
+  function getActionLabel() {
+    if (typeof label === "string") {
+      return label;
+    }
+
+    if (activeItem) {
+      return label(activeItem);
+    }
+  }
+
   return (
     <button type="button" className={styles.action} onClick={handleClick}>
       <Typography textColor={color}>
-        <span className={styles.label}>{label}</span>
+        <span className={styles.label}>{getActionLabel()}</span>
       </Typography>
 
       {icon && <Icon name={icon} color={color} />}