+++ packages/components-native/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.51.0](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.50.0...@jobber/components-native@0.51.0) (2023-12-05)
+
+
+### Features
+
+* **components-native:** Add stepped variation of progress bar for mobile ([#1669](https://github.com/GetJobber/atlantis/issues/1669)) ([d3095f7](https://github.com/GetJobber/atlantis/commit/d3095f7261ff1b3c5005b64dda6f61ac21fc0b95))
+
+
+
+
+
+# [0.50.0](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.49.2...@jobber/components-native@0.50.0) (2023-12-05)
+
+
+### Features
+
+* **hooks:** Move useShowClear to @jobber/hooks ([#1675](https://github.com/GetJobber/atlantis/issues/1675)) ([7764bb5](https://github.com/GetJobber/atlantis/commit/7764bb5a68b9ff8e43c5dc386d3829f4c8e46160))
+
+
+
+
+
+## [0.49.2](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.49.1...@jobber/components-native@0.49.2) (2023-12-05)
+
+**Note:** Version bump only for package @jobber/components-native
+
+
+
+
+
+## [0.49.1](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.49.0...@jobber/components-native@0.49.1) (2023-12-01)
+
+**Note:** Version bump only for package @jobber/components-native
+
+
+
+
+
 # [0.49.0](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.48.11...@jobber/components-native@0.49.0) (2023-11-29)