+++ packages/components/CHANGELOG.md
@@ -3,6 +3,118 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.67.5](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.67.4...@jobber/components@4.67.5) (2023-12-08)
+
+
+### Bug Fixes
+
+* **components:** Clean up interactive state colors in combobox ([#1679](https://github.com/GetJobber/atlantis/issues/1679)) ([9aa2433](https://github.com/GetJobber/atlantis/commit/9aa243369b8b14be58a1da9e792d3cc754aa10b3))
+* **components:** update spacing of DataList bulk header actions ([#1676](https://github.com/GetJobber/atlantis/issues/1676)) ([080c07c](https://github.com/GetJobber/atlantis/commit/080c07cb01bf6842de106c51f167120ec2a168a9))
+
+
+
+
+
+## [4.67.4](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.67.3...@jobber/components@4.67.4) (2023-12-08)
+
+
+### Bug Fixes
+
+* **components:** Fix DataList tags counter overflow ([#1680](https://github.com/GetJobber/atlantis/issues/1680)) ([ce99e57](https://github.com/GetJobber/atlantis/commit/ce99e57f3dcb66bc3e63feb2fab98d5036635e10))
+
+
+
+
+
+## [4.67.3](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.67.2...@jobber/components@4.67.3) (2023-12-08)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.67.2](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.67.1...@jobber/components@4.67.2) (2023-12-08)
+
+
+### Bug Fixes
+
+* **components:** remove gradient on scroll from top and bottom of Combobox [JOB-79443] ([#1678](https://github.com/GetJobber/atlantis/issues/1678)) ([0e56148](https://github.com/GetJobber/atlantis/commit/0e561480c26ae3bb9b3c6e349ec7219245e235e1))
+
+
+
+
+
+## [4.67.1](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.67.0...@jobber/components@4.67.1) (2023-12-05)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+# [4.67.0](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.66.7...@jobber/components@4.67.0) (2023-12-05)
+
+
+### Features
+
+* **components:** expose search term on combobox action ([#1673](https://github.com/GetJobber/atlantis/issues/1673)) ([ebaa01f](https://github.com/GetJobber/atlantis/commit/ebaa01f781cc5cfb53b6f2944f772bb9c4f40086))
+
+
+
+
+
+## [4.66.7](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.66.6...@jobber/components@4.66.7) (2023-12-05)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.66.6](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.66.5...@jobber/components@4.66.6) (2023-12-01)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.66.5](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.66.4...@jobber/components@4.66.5) (2023-12-01)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.66.4](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.66.3...@jobber/components@4.66.4) (2023-11-30)
+
+
+### Bug Fixes
+
+* **components:** multiselect preventing space bar from being recognized ([#1665](https://github.com/GetJobber/atlantis/issues/1665)) ([f920fda](https://github.com/GetJobber/atlantis/commit/f920fda1d4b60fe1fea0be2785b9bf327393d3e2))
+
+
+
+
+
+## [4.66.3](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.66.2...@jobber/components@4.66.3) (2023-11-30)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.66.2](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.66.1...@jobber/components@4.66.2) (2023-11-30)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
 ## [4.66.1](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.66.0...@jobber/components@4.66.1) (2023-11-30)